Add fight simulation and feature for requesting army improve.

+++ app/src/main/java/com/example/gossipwars/logic/entities/Army.kt
@@ -1,7 +1,7 @@
 package com.example.gossipwars.logic.entities
 
-data class Army(var size : Int = 100000, var attack : Double = 50.0,
-                var defense : Double = 50.0) {
+data class Army(var size : Int = 100000, var attack : Int = 50,
+                var defense : Int = 50) {
     var sizePerRegion : MutableMap<Int, Int> = mutableMapOf()
 
     companion object Factory {
